Issue #17 - Refactor code from "utils.js" (and add ESLint) (#23)
* Issue #17 - Move constants and miscellaneous functions from utils to separate files

* Issue #17 - Move storage and local store from utils to separated files

* Issue #17 - Move provider helper from utils to his own folder

* Issue #17 - Move config file from utils to own file (and fix minor bugs)

* Issue #17 - Add ESList, fix Lint warnings and re-declare some dependencies as standard (instead of dev)

* Changed branding for CRITICALSTART on many pages. Also some minor branding changes to be consistent on how we reference VirusTotal and a few other minor changes. Most links now point to the open source GitHub repo instead of our website.

* Update minor version

* Issue #17 - Fix import for Gibberish AES
