+++ package.json
@@ -7,25 +7,30 @@
   "private": true,
   "scripts": {
     "build": "webpack --config webpack.config.js",
+    "lint": "eslint --ext js src",
+    "lint:fix": "eslint --ext js src --fix",
     "live": "webpack --config webpack.config.js --watch"
   },
   "devDependencies": {
-    "@fortawesome/fontawesome-free": "^5.15.3",
-    "bootstrap": "^4.6.0",
     "clean-webpack-plugin": "^4.0.0-alpha.0",
     "copy-webpack-plugin": "^8.1.1",
     "css-loader": "^5.2.4",
+    "eslint": "^7.25.0",
     "file-loader": "^6.2.0",
-    "gibberish-aes": "^1.0.0",
     "html-minimizer-webpack-plugin": "^2.1.0",
+    "style-loader": "^2.0.0",
+    "webpack": "^5.36.2",
+    "webpack-cli": "^4.6.0"
+  },
+  "dependencies": {
+    "@fortawesome/fontawesome-free": "^5.15.3",
+    "bootstrap": "^5.0.0",
+    "gibberish-aes": "^1.0.0",
     "js-beautify": "^1.13.13",
     "lodash": "^4.17.21",
     "luxon": "^1.26.0",
     "mustache": "^4.2.0",
     "notiflix": "^2.7.0",
-    "sortablejs": "^1.13.0",
-    "style-loader": "^2.0.0",
-    "webpack": "^5.36.2",
-    "webpack-cli": "^4.6.0"
+    "sortablejs": "^1.13.0"
   }
 }
